chore: move AuthWrapper out to common
