+++ src/pages/Research/Content/ResearchList.tsx
@@ -1,6 +1,6 @@
 import { observer } from 'mobx-react'
 import { Box, Flex, Heading } from 'theme-ui'
-import { AuthWrapper } from 'src/components/Auth/AuthWrapper'
+import { AuthWrapper } from 'src/common/AuthWrapper'
 import { Button } from 'oa-components'
 import ResearchListItem from './ResearchListItem'
 import { useResearchStore } from 'src/stores/Research/research.store'
