+++ src/pages/admin/Admin.tsx
@@ -5,7 +5,7 @@ import { AdminTags } from './content/AdminTags'
 import { AdminUsers } from './content/AdminUsers'
 import { AdminCategories } from './content/AdminCategories'
 import { AdminBetaTesters } from './content/AdminBetaTesters'
-import { AuthWrapper } from 'src/components/Auth/AuthWrapper'
+import { AuthWrapper } from 'src/common/AuthWrapper'
 import { Flex, Box, Text } from 'theme-ui'
 import { Link } from 'react-router-dom'
 import { inject, observer } from 'mobx-react'
