+++ src/pages/common/Header/Menu/MenuMobile/MenuMobilePanel.tsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled'
 import { Box } from 'theme-ui'
 import Profile from 'src/pages/common/Header/Menu/Profile/Profile'
 import MenuMobileLink from 'src/pages/common/Header/Menu/MenuMobile/MenuMobileLink'
-import { AuthWrapper } from 'src/components/Auth/AuthWrapper'
+import { AuthWrapper } from 'src/common/AuthWrapper'
 import { getSupportedModules } from 'src/modules'
 import { inject } from 'mobx-react'
 import type { ThemeStore } from 'src/stores/Theme/theme.store'
